requirements.txt
PyGithub
pandas
seaborn
matplotlib